+++ macros/sql/get_relations_by_pattern.sql
@@ -15,7 +15,12 @@
     {%- if table_list and table_list['table'] -%}
         {%- set tbl_relations = [] -%}
         {%- for row in table_list['table'] -%}
-            {%- set tbl_relation = api.Relation.create(database, row.table_schema, row.table_name) -%}
+            {%- set tbl_relation = api.Relation.create(
+                database=database,
+                schema=row.table_schema,
+                identifier=row.table_name,
+                type=row.table_type
+            ) -%}
             {%- do tbl_relations.append(tbl_relation) -%}
         {%- endfor -%}